added validators

+++ src/services/Posts/controllers/deletePost.ts
@@ -1,9 +1,17 @@
 import PostsModel from '../model';
 import type { Response } from '../..';
 import type { getPostById } from '../routes';
+import { validatePostParams } from '../validators/validatePostParams';
 
 // Delete by ID
 const deletePost = async (req: getPostById, res: Response): Promise<void> => {
+  // Validate params of the post
+  if (!(await validatePostParams(req.params))) {
+    return res
+      .status(400)
+      .json({ message: 'Invalid guid for deletion', data: req.params });
+  }
+
   try {
     // Find the document by ID and delete it
     const deletedPost = await PostsModel.deleteMany({ guid: req.params.guid });
